refactor: model recording backends
