+++ src/daemon/handlers/record-trace-recording.ts
@@ -21,41 +21,19 @@ import {
   RECORDING_EXPORT_QUALITIES,
   recordingQualityInputToExportQuality,
 } from '../../core/recording-export-quality.ts';
-import {
-  buildRecordStopFailure,
-  formatRecordTraceError,
-  formatRecordTraceExecFailure,
-} from '../record-trace-errors.ts';
 import { resolveRecordingProvider } from '../recording-provider.ts';
-import { finalizeRecordingOverlay } from './record-trace-finalize.ts';
 import { errorResponse } from './response.ts';
-import { startAndroidRecording, stopAndroidRecording } from './record-trace-android.ts';
 import { deriveAndroidChunkOutPath } from './record-trace-android-chunks.ts';
-import { withDiagnosticTimer } from '../../utils/diagnostics.ts';
 import {
-  getIosRunnerOptions,
-  normalizeAppBundleId,
-  startIosDeviceRecording,
-  startMacOsRecording,
-  stopMacOsRecording,
-  stopIosDeviceRecording,
-  warmIosSimulatorRunner,
-} from './record-trace-ios.ts';
-import {
-  IOS_SIMULATOR_RECORDING_STOP_TIMEOUT_MS,
-  stopIosSimulatorRecordingProcess,
-} from './record-trace-ios-simulator.ts';
+  resolveRecordingBackendForDevice,
+  resolveRecordingBackendForRecording,
+} from './record-trace-recording-backends.ts';
+import type { RecordTraceDeps, RecordingBase } from './record-trace-types.ts';
 import { resolveImplicitSessionScope, resolvePublicSessionName } from '../session-routing.ts';
 
 const IOS_DEVICE_RECORD_MIN_FPS = 1;
 const IOS_DEVICE_RECORD_MAX_FPS = 120;
-const LOCAL_RECORDING_READY_POLL_MS = 250;
-const LOCAL_RECORDING_READY_SETTLE_POLLS = 2;
 const IOS_SIMULATOR_RECORDING_TAIL_SETTLE_MS = 350;
-const IOS_SIMULATOR_VIDEO_READY_POLL_MS = 150;
-const IOS_SIMULATOR_VIDEO_READY_ATTEMPTS = 12;
-
-import type { RecordTraceDeps, RecordingBase } from './record-trace-types.ts';
 
 export type { RecordTraceDeps, RecordingBase } from './record-trace-types.ts';
 
@@ -96,93 +74,6 @@ function buildRecordingBase(req: DaemonRequest, outPath: string): RecordingBase 
   };
 }
 
-async function waitForLocalRecordingSettleWindow(outPath: string): Promise<number> {
-  // simctl recordVideo can take a beat to open its output even though recording has already
-  // started. This is a short settle window, not a strict readiness guarantee. We prefer a
-  // close recorder anchor over blocking start indefinitely waiting for non-zero bytes.
-  for (let attempt = 0; attempt < LOCAL_RECORDING_READY_SETTLE_POLLS; attempt += 1) {
-    try {
-      const stat = fs.statSync(outPath);
-      if (stat.size > 0) {
-        return Date.now();
-      }
-    } catch {
-      // Wait for the recorder to create the output file.
-    }
-
-    if (attempt + 1 >= LOCAL_RECORDING_READY_SETTLE_POLLS) {
-      return Date.now();
-    }
-
-    await sleep(LOCAL_RECORDING_READY_POLL_MS);
-  }
-
-  return Date.now();
-}
-
-// --- Per-platform start helpers ---
-
-async function startIosSimulatorRecording(params: {
-  req: DaemonRequest;
-  activeSession: SessionState;
-  device: SessionState['device'];
-  logPath?: string;
-  deps: RecordTraceDeps;
-  recordingBase: RecordingBase;
-  resolvedOut: string;
-}): Promise<DaemonResponse | NonNullable<SessionState['recording']>> {
-  const { req, activeSession, device, logPath, deps, recordingBase, resolvedOut } = params;
-
-  // The warm-up carries the gesture-clock anchor on its snapshot response when the runner
-  // stamps it, letting us skip a standalone uptime command. The anchor is a pure clock pair
-  // (origin uptime + daemon receipt time), so capturing it before the recorder spawn/settle
-  // window is equivalent to capturing it after: recordingStartedAt stays readyAt below.
-  const warmAnchor = recordingBase.showTouches
-    ? await warmIosSimulatorRunner({ req, activeSession, device, logPath, deps })
-    : undefined;
-  const { child, wait } = deps.startIosSimulatorRecording({ device, outPath: resolvedOut });
-  const readyAt = await waitForLocalRecordingSettleWindow(resolvedOut);
-  let gestureClockOriginAtMs: number | undefined;
-  let gestureClockOriginUptimeMs: number | undefined;
-  if (warmAnchor) {
-    gestureClockOriginAtMs = warmAnchor.gestureClockOriginAtMs;
-    gestureClockOriginUptimeMs = warmAnchor.gestureClockOriginUptimeMs;
-  } else if (recordingBase.showTouches) {
-    // Fallback for older runner builds (or a failed/unavailable warm anchor): issue a
-    // standalone uptime command and pair it at the request midpoint.
-    try {
-      const uptimeRequestStartedAtMs = Date.now();
-      const uptimeResult = await deps.runIosRunnerCommand(
-        device,
-        {
-          command: 'uptime',
-          appBundleId: normalizeAppBundleId(activeSession),
-        },
-        getIosRunnerOptions(req, logPath, activeSession),
-      );
-      const uptimeRequestFinishedAtMs = Date.now();
-      gestureClockOriginAtMs = Math.round(
-        (uptimeRequestStartedAtMs + uptimeRequestFinishedAtMs) / 2,
-      );
-      gestureClockOriginUptimeMs =
-        typeof uptimeResult.currentUptimeMs === 'number' ? uptimeResult.currentUptimeMs : undefined;
-    } catch {
-      // Best effort only; wall-clock fallback remains available.
-    }
-  }
-  return {
-    platform: 'ios',
-    child,
-    wait,
-    ...recordingBase,
-    recorderPid: child.pid,
-    startedAt: readyAt,
-    gestureClockOriginAtMs:
-      gestureClockOriginUptimeMs === undefined ? undefined : gestureClockOriginAtMs,
-    gestureClockOriginUptimeMs,
-  };
-}
-
 // --- Start recording orchestrator ---
 
 // fallow-ignore-next-line complexity
@@ -204,6 +95,11 @@ async function startRecording(params: {
   const fpsFlag = req.flags?.fps;
   const qualityFlag = req.flags?.quality;
   const maxSizeFlag = req.flags?.screenshotMaxSize;
+  const backend = resolveRecordingBackendForDevice(device);
+  const platformValidationError = backend.validateStart?.(req) ?? null;
+  if (platformValidationError) {
+    return platformValidationError;
+  }
   if (
     fpsFlag !== undefined &&
     (!Number.isInteger(fpsFlag) ||
@@ -231,63 +127,27 @@ async function startRecording(params: {
     return errorResponse('UNSUPPORTED_OPERATION', 'record is not supported on this device');
   }
 
-  const outPath = req.positionals?.[1] ?? `./recording-${Date.now()}.mp4`;
+  const outputPath = backend.resolveOutputPath({ req, device });
+  if (!outputPath.ok) {
+    return outputPath.response;
+  }
+  const outPath = outputPath.path;
   const resolvedOut = SessionStore.expandHome(outPath, req.meta?.cwd);
   const recordingBase = buildRecordingBase(req, resolvedOut);
   fs.mkdirSync(path.dirname(resolvedOut), { recursive: true });
   fs.rmSync(resolvedOut, { force: true });
 
-  let recording: NonNullable<SessionState['recording']> | DaemonResponse;
-  if (device.platform === 'ios' && device.kind === 'device') {
-    const appBundleId = normalizeAppBundleId(activeSession);
-    if (!appBundleId) {
-      return errorResponse(
-        'INVALID_ARGS',
-        'record on physical iOS devices requires an active app session; run open <app> first',
-      );
-    }
-    recording = await startIosDeviceRecording({
-      req,
-      activeSession,
-      sessionStore,
-      device,
-      logPath,
-      deps,
-      fpsFlag,
-      recordingBase,
-      appBundleId,
-    });
-  } else if (device.platform === 'macos') {
-    const appBundleId = normalizeAppBundleId(activeSession);
-    if (!appBundleId) {
-      return errorResponse(
-        'INVALID_ARGS',
-        'record on macOS requires an active app session; run open <app> first',
-      );
-    }
-    recording = await startMacOsRecording({
-      req,
-      activeSession,
-      device,
-      logPath,
-      deps,
-      fpsFlag,
-      recordingBase,
-      appBundleId,
-    });
-  } else if (device.platform === 'ios') {
-    recording = await startIosSimulatorRecording({
-      req,
-      activeSession,
-      device,
-      logPath,
-      deps,
-      recordingBase,
-      resolvedOut,
-    });
-  } else {
-    recording = await startAndroidRecording({ device, recordingBase });
-  }
+  const recording = await backend.start({
+    req,
+    activeSession,
+    sessionStore,
+    device,
+    logPath,
+    deps,
+    fpsFlag,
+    recordingBase,
+    resolvedOut,
+  });
 
   if ('ok' in recording) {
     return recording;
@@ -314,128 +174,6 @@ async function startRecording(params: {
   };
 }
 
-// --- Stop recording helpers ---
-
-async function stopNonRunnerRecording(params: {
-  deps: RecordTraceDeps;
-  device: SessionState['device'];
-  recording: Extract<NonNullable<SessionState['recording']>, { platform: 'ios' | 'android' }>;
-  stopRequestedAt: number;
-}): Promise<DaemonResponse | null> {
-  const { deps, device, recording, stopRequestedAt } = params;
-  if (recording.platform === 'android') {
-    return await stopAndroidRecording({ deps, device, recording, stopRequestedAt });
-  }
-
-  await withDiagnosticTimer('record_stop_tail_settle', () => deps.waitForRecordingTail(recording), {
-    platform: recording.platform,
-    gestureEventCount: recording.gestureEvents.length,
-  });
-  const stopResult = await withDiagnosticTimer(
-    'record_stop_ios_simulator_process',
-    () => stopIosSimulatorRecordingProcess({ deps, recording }),
-    {
-      outPath: recording.outPath,
-    },
-  );
-  if (!stopResult) {
-    return buildIosSimulatorRecordingStopFailure(
-      `failed to stop recording: simctl recordVideo did not exit after ${IOS_SIMULATOR_RECORDING_STOP_TIMEOUT_MS}ms and forced cleanup`,
-      recording,
-      stopRequestedAt,
-    );
-  }
-  if (stopResult.exitCode !== 0) {
-    return buildIosSimulatorRecordingStopFailure(
-      `failed to stop recording: ${formatRecordTraceExecFailure(stopResult, 'simctl recordVideo')}`,
-      recording,
-      stopRequestedAt,
-    );
-  }
-
-  await withDiagnosticTimer(
-    'record_stop_video_stable',
-    () =>
-      deps.waitForStableFile(recording.outPath, {
-        pollMs: IOS_SIMULATOR_VIDEO_READY_POLL_MS,
-        attempts: IOS_SIMULATOR_VIDEO_READY_ATTEMPTS,
-      }),
-    {
-      outPath: recording.outPath,
-    },
-  );
-  const playable = await withDiagnosticTimer(
-    'record_stop_video_playable_check',
-    () => deps.isPlayableVideo(recording.outPath),
-    {
-      outPath: recording.outPath,
-    },
-  );
-  if (!playable) {
-    return buildIosSimulatorRecordingStopFailure(
-      `failed to stop recording: ${recording.outPath} was not finalized into a playable video`,
-      recording,
-      stopRequestedAt,
-    );
-  }
-
-  if (recording.maxSize !== undefined) {
-    try {
-      await withDiagnosticTimer(
-        'record_stop_resize',
-        () =>
-          deps.resizeRecording({
-            videoPath: recording.outPath,
-            maxSize: recording.maxSize!,
-            exportQuality: recording.exportQuality,
-            targetLabel: 'iOS recording',
-          }),
-        {
-          outPath: recording.outPath,
-          maxSize: recording.maxSize,
-        },
-      );
-    } catch (error) {
-      recording.overlayWarning = `failed to resize recording: ${formatRecordTraceError(error)}`;
-    }
-  }
-
-  await withDiagnosticTimer(
-    'record_stop_finalize_overlay',
-    () =>
-      finalizeRecordingOverlay({
-        recording,
-        deps,
-        targetLabel: 'iOS recording',
-      }),
-    {
-      outPath: recording.outPath,
-      showTouches: recording.showTouches,
-      gestureEventCount: recording.gestureEvents.length,
-    },
-  );
-
-  return null;
-}
-
-function buildIosSimulatorRecordingStopFailure(
-  message: string,
-  recording: Extract<NonNullable<SessionState['recording']>, { platform: 'ios' }>,
-  stopRequestedAt: number,
-): DaemonResponse {
-  const failure = buildRecordStopFailure(message, recording, stopRequestedAt);
-  removeInvalidRecordingOutput(recording.outPath);
-  return errorResponse('COMMAND_FAILED', failure.message);
-}
-
-function removeInvalidRecordingOutput(outPath: string): void {
-  try {
-    fs.rmSync(outPath, { force: true });
-  } catch {
-    // Best effort: the error response still reports the failed finalization.
-  }
-}
-
 async function stopRecording(params: {
   req: DaemonRequest;
   activeSession: SessionState;
@@ -453,13 +191,17 @@ async function stopRecording(params: {
   const stopRequestedAt = Date.now();
   const invalidatedReason = recording.invalidatedReason;
   activeSession.recording = undefined;
+  const backend = resolveRecordingBackendForRecording(recording);
 
-  const stopError =
-    recording.platform === 'ios-device-runner'
-      ? await stopIosDeviceRecording({ req, activeSession, device, logPath, deps, recording })
-      : recording.platform === 'macos-runner'
-        ? await stopMacOsRecording({ req, activeSession, device, logPath, deps, recording })
-        : await stopNonRunnerRecording({ deps, device, recording, stopRequestedAt });
+  const stopError = await backend.stop({
+    req,
+    activeSession,
+    device,
+    logPath,
+    deps,
+    recording,
+    stopRequestedAt,
+  });
   if (stopError) {
     return stopError;
   }
@@ -541,15 +283,17 @@ function deriveClientTelemetryPath(
   return deriveRecordingTelemetryPath(recording.clientOutPath);
 }
 
-function releaseRecordOnlySession(
+async function releaseRecordOnlySession(
   sessionStore: SessionStore,
   sessionName: string,
   session: SessionState,
   options: { writeLog?: boolean } = {},
-): void {
+): Promise<void> {
   if (!session.recordOnlySession) {
     return;
   }
+  const backend = resolveRecordingBackendForDevice(session.device);
+  await backend.cleanupRecordOnlySession?.(session);
   if (options.writeLog) {
     sessionStore.writeSessionLog(session);
   }
@@ -594,7 +338,7 @@ export async function handleRecordCommand(params: {
 
   const response = await stopRecording({ req, activeSession, device, logPath, deps });
   if (!response.ok) {
-    releaseRecordOnlySession(sessionStore, sessionName, activeSession);
+    await releaseRecordOnlySession(sessionStore, sessionName, activeSession);
     return response;
   }
 
@@ -608,6 +352,6 @@ export async function handleRecordCommand(params: {
       showTouches: response.data?.showTouches,
     },
   });
-  releaseRecordOnlySession(sessionStore, sessionName, activeSession, { writeLog: true });
+  await releaseRecordOnlySession(sessionStore, sessionName, activeSession, { writeLog: true });
   return response;
 }
